Create and update a scenes.json file in scenesDir
This file could be used by the game to load a list of available
scenes without using php. The scenes.json file is automatically
updated everytime the editor queues the list of scenes (that means
at least everytime you reload the editor).

You should not need to manage the scenes.json file on your own!

References #9